Move data sync code to the sync directory.
