+++ sdk/src/Platforms/cordova/MobileServiceSqliteStore.js
@@ -11,7 +11,7 @@ var Platform = require('Platforms/Platform'),
     Validate = require('../../Utilities/Validate'),
     _ = require('../../Utilities/Extensions'),
     queryHelper = require('azure-mobile-apps/src/query'),
-    ColumnType = require('./ColumnType'),
+    ColumnType = require('../../sync/ColumnType'),
     sqliteSerializer = require('./sqliteSerializer'),
     Query = require('query.js').Query,
     formatSql = require('azure-odata-sql').format;
