+++ sdk/src/Platforms/cordova/sqliteSerializer.js
@@ -11,7 +11,7 @@
 var Platform = require('Platforms/Platform'),
     Validate = require('../../Utilities/Validate'),
     _ = require('../../Utilities/Extensions'),
-    ColumnType = require('./ColumnType'),
+    ColumnType = require('../../sync/ColumnType'),
     verror = require('verror'),
     typeConverter = require('./typeConverter');
