+++ sdk/src/sync/MobileServiceSyncTable.js
@@ -2,7 +2,7 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // ----------------------------------------------------------------------------
 
-var Validate = require('./Utilities/Validate');
+var Validate = require('../Utilities/Validate');
 var Platform = require('Platforms/Platform');
 
 /**
